+++ tests/ExpressionStringEvaluator.Tests/IntegrationTests.cs
@@ -210,7 +210,7 @@ public sealed class IntegrationTests : IDisposable
     [InlineData("-f \"{Repository.Location} {Repository.Name}\"", "-f \"D:\\repositories ExpressionStringEvaluator\"")]
     [InlineData("-f \"{Repository.Location}\\\\{Repository.Name}\"", "-f \"D:\\repositories\\ExpressionStringEvaluator\"")]
     [InlineData("-f \"{Repository.Location}{backslash}{Repository.Name}\"", "-f \"D:\\repositories\\ExpressionStringEvaluator\"")]
-    // [InlineData("-f \"{Repository.Location}{PathSeparator}{Repository.Name}\"", "-f \"D:\\repositories\\ExpressionStringEvaluator\"")] // depends on system
+    /*[InlineData("-f \"{Repository.Location}{PathSeparator}{Repository.Name}\"", "-f \"D:\\repositories\\ExpressionStringEvaluator\"")] // depends on system*/
     public void ParseRepoContext(string input, string expectedOutput)
     {
         // arrange
@@ -295,125 +295,5 @@ public sealed class IntegrationTests : IDisposable
 
     public class Context
     {
-    }
-}
-
-
-public class Repository
-{
-    public string Name { get; set; }
-
-    public string Path { get; set; }
-
-    public string Location { get; set; }
-
-    public string CurrentBranch { get; set; }
-
-    public string[] Branches { get; set; }
-
-    public string[] LocalBranches { get; set; }
-
-    public bool CurrentBranchHasUpstream { get; set; }
-
-    public bool CurrentBranchIsDetached { get; set; }
-
-    public bool CurrentBranchIsOnTag { get; set; }
-
-    public int? AheadBy { get; set; }
-
-    public int? BehindBy { get; set; }
-
-    public int? LocalUntracked { get; set; }
-
-    public int? LocalModified { get; set; }
-
-    public int? LocalMissing { get; set; }
-
-    public int? LocalAdded { get; set; }
-
-    public int? LocalStaged { get; set; }
-
-    public int? LocalRemoved { get; set; }
-
-    public int? LocalIgnored { get; set; }
-
-    public int? StashCount { get; set; }
-
-    public string[] RemoteUrls { get; set; }
-
-    public string SafePath
-    {
-        // use '/' for linux systems and bash command line (will work on cmd and powershell as well)
-        get
-        {
-            var safePath = Path?.Replace(@"\", "/") ?? string.Empty;
-            if (safePath.EndsWith("/"))
-            {
-                safePath = safePath.Substring(0, safePath.Length - 1);
-            }
-
-            return safePath;
-        }
-    }
-}
-
-public class RepositoryVariableProvider : IVariableProvider<Repository>
-{
-    public bool CanProvide(string key)
-    {
-        return !string.IsNullOrWhiteSpace(key) && key.StartsWith("Repository.", StringComparison.CurrentCultureIgnoreCase);
-    }
-
-    public string Provide(Repository context, string key, string arg)
-    {
-        var startIndex = "Repository.".Length;
-        var k = key.Substring(startIndex, key.Length - startIndex);
-
-        if ("Name".Equals(k, StringComparison.CurrentCultureIgnoreCase))
-        {
-            return context.Name;
-        }
-
-        if ("Path".Equals(k, StringComparison.CurrentCultureIgnoreCase))
-        {
-            return context.Path;
-        }
-
-        if ("SafePath".Equals(k, StringComparison.CurrentCultureIgnoreCase))
-        {
-            return context.SafePath;
-        }
-
-        if ("Location".Equals(k, StringComparison.CurrentCultureIgnoreCase))
-        {
-            return context.Location;
-        }
-
-        if ("CurrentBranch".Equals(k, StringComparison.CurrentCultureIgnoreCase))
-        {
-            return context.CurrentBranch;
-        }
-
-        if ("Branches".Equals(k, StringComparison.CurrentCultureIgnoreCase))
-        {
-            return string.Join("|", context.Branches);
-        }
-
-        if ("LocalBranches".Equals(k, StringComparison.CurrentCultureIgnoreCase))
-        {
-            return string.Join("|", context.LocalBranches);
-        }
-
-        if ("RemoteUrls".Equals(k, StringComparison.CurrentCultureIgnoreCase))
-        {
-            return string.Join("|", context.RemoteUrls);
-        }
-
-        throw new NotImplementedException();
-    }
-
-    public string Provide(string key, string arg)
-    {
-        throw new NotImplementedException();
     }
 }
